challenge-2/requirements.txt
numpy==1.24.4
pandas==2.0.3
matplotlib==3.7.1
seaborn==0.12.2
tqdm==4.65.0
Pillow==9.5.0
opencv-python==4.7.0.72
shutilwhich 
torch==2.0.1
torchvision==0.15.2
ipykernel==6.23.1
notebook==6.5.4